% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postop.R
\docType{data}
\name{postop}
\alias{postop}
\title{Transmitter operated mice}
\format{A data frame with post-op data.}
\usage{
data(postop)
}
\description{
Data from a transmitter implantation experiment with various behavioural data. Also includes sham-operated animals.
Observation time was over 30 days, starting with day=-1 as baseline (physiological state). The treatment columns shows
what operational state the animal is in and the condition field shows two possible types of analgesia.
}
\examples{
data(postop)
head(postop)
}
\keyword{datasets}
