% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsa.R
\name{relsa}
\alias{relsa}
\title{Relative Severity Assessment (RELSA) Score}
\usage{
relsa(set, bsl, a = 1, drop = NULL, turnvars = NULL, relsaNA = NA)
}
\arguments{
\item{set}{dataset sample data frame}

\item{bsl}{reference or baseline data (must be in RELSA format, see relsa_load)}

\item{a}{unique animal index in data set}

\item{drop}{character vector of variables to drop}

\item{turnvars}{variables with inverted development under severity (i.e. temperature might rise or drop depending on the actual animal model)}

\item{relsaNA}{codes how to handle NaN values during calculations (default is NA)}
}
\value{
\code{relsa} list with RELSA results (differences and weights plus RELSA score)
}
\description{
The \code{relsa} function calculates a composite relative severity score based on normalized
differences. Further, the values are regularized with a reference set to estimate relative
severity.
}
\details{
\cr PLEASE NOTE: Changing the reference set will change the relative severity context.
}
\examples{
\dontrun{
RELSA <- relsa(set=testset, bsl, a=1, drop=c("bw","score"), turnvars="temp")
}

}
