% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsa_levels.R
\name{relsa_levels}
\alias{relsa_levels}
\title{k+1 levels determination for RELSA reference data}
\usage{
relsa_levels(refset, mypath, filename, bsl, drops = NULL, turns = NULL,
  relsaNA = NA, k = 4, showScree = "no", customCol = NULL,
  seed = 123, myYlim = c(0, 1.4), saveTiff = "yes")
}
\arguments{
\item{refset}{normalized data from a reference set}

\item{mypath}{path specification for saving clustering outputs}

\item{filename}{filename for the output *.tiff}

\item{bsl}{baseline object from relsa_baselines function}

\item{drops}{character vector of variables to drop}

\item{turns}{variables with inverted development under severity (i.e. temperature might rise or drop depending on the actual animal model)}

\item{relsaNA}{codes how to handle NaN values during calculations (default is NA)}

\item{k}{the number of k clusters (user defined, default is 4)}

\item{showScree}{show the Scree plot to estimate the optimal number of clusters}

\item{customCol}{custom color string for the k clusters}

\item{seed}{random seed, if set to NULL, no seeding for kmeans is used}

\item{myYlim}{limits of the y-axis for the clustering}

\item{saveTiff}{yes/no whether to save the output as *.tiff file (if no, mypath is obsolete)}
}
\value{
\code{levels} data.frame with k+1 levels determined by k-means
}
\description{
The \code{relsa_levels} function calculates kmeans-based RELSA levels based on a reference data set.
}
