% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsa_load.R
\name{relsa_load}
\alias{relsa_load}
\title{Loads data in RELSA Format}
\usage{
relsa_load(file, treatment = NULL, condition = NULL)
}
\arguments{
\item{file}{dataset sample data frame}

\item{treatment}{treatment string (e.g., "Transmitter")}

\item{condition}{condition string (e.g. Carprofen) in the sample table}
}
\value{
\code{set} data.frame with set sample data
}
\description{
The \code{relsa_load} function loads data into a ready-to-use format for RELSA.
}
