% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsa_norm.R
\name{relsa_norm}
\alias{relsa_norm}
\title{Data normalization for RELSA score calculation}
\usage{
relsa_norm(set, normthese = NULL, ontime = 1)
}
\arguments{
\item{set}{dataset sample data frame}

\item{normthese}{variable string in the input data set that shall be normalized}

\item{ontime}{unique reference time point in dataset (e.g. 1 for day 1)}
}
\value{
\code{set} data.frame with normalized column
}
\description{
The \code{relsa_norm} function normalizes defined variables to 100 % at a given time point.
}
