% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsa_plot.R
\name{relsa_plot}
\alias{relsa_plot}
\title{Plot RELSA score and variables together}
\usage{
relsa_plot(
  set = NULL,
  RELSA,
  levels = NULL,
  animal = 1,
  plotvar = NULL,
  plotRELSA = TRUE,
  myylim = c(70, 110),
  myYlim = c(0, 2),
  mypch = 1,
  mycol = "red",
  myXlab = "time point"
)
}
\arguments{
\item{set}{dataset sample data frame}

\item{RELSA}{RELSA object}

\item{levels}{k+1 levels from kmeans calculations (relsa_levels function)}

\item{animal}{which unique animal in the set?}

\item{plotvar}{which variable shall be plotted (does not scale with RELSA score)}

\item{plotRELSA}{plot the RELSA score "yes/no"?}

\item{myylim}{custom y-axis limits (left axis)}

\item{myYlim}{custom y-axis limits (right axis)}

\item{mypch}{custom dot marker}

\item{mycol}{custom plot color}

\item{myXlab}{custom x-axis label}
}
\value{
\code{relsa} data.frame with RELSA scores for each variable, the sum and mean score
}
\description{
The \code{relsa_plot} function visualizes the RELSA score alone or in combination with one of the variables.
}
