% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsa_reference.R
\name{relsa_reference}
\alias{relsa_reference}
\title{RELSA Default Reference Model}
\usage{
relsa_reference(
  vars = c("bwc", "burON", "hr", "hrv", "temp", "act"),
  normthese = c("hr", "burON", "hrv", "temp", "act"),
  turnthese = c("hr", "temp")
)
}
\value{
list with baselines (bsl) and cluster levels (levels)
}
\description{
The \code{relsa_reference} functions calculates the default RELSA reference model - to keeps things simple.
The reference set is the surgery data (transmitter-implanted, female B6 mice)
}
