% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsa_wrapper.R
\name{relsa_wrapper}
\alias{relsa_wrapper}
\title{RELSA Wrapper}
\usage{
relsa_wrapper(
  querydata,
  baseline = NULL,
  treatment = NULL,
  condition = NULL,
  normthese = NULL,
  turnsQuery = NULL,
  dropsQuery = NULL,
  animalnr = 1,
  ymax = 1.2,
  ymin = 0,
  pcadims = 2,
  studylabel = NULL,
  severity = NULL,
  colorlabel = NULL,
  showScree = "yes",
  showPlot = "no",
  k = 6
)
}
\arguments{
\item{querydata}{dataset sample data frame}

\item{baseline}{cluster baselines from \code{relsa_baselines} function}

\item{treatment}{treatment column, potential data filter for subsampling (use column name as character, e.g. "transmitter")}

\item{condition}{condition column, potential data filter for subsampling (use column name as character, e.g. "Carprofen")}

\item{normthese}{variable vector (colum names as character, e.g. c("hr","temp")) with the names of the variables that need "normalization"}

\item{turnsQuery}{variable vector (colum names as character, e.g. c("hr","temp")) with the names of the variables that need "turning"}

\item{dropsQuery}{variables to drop}

\item{animalnr}{animal number (not id) in the data set for which an example RELSA flow is generated (defaults to 1)}

\item{ymax}{y-max of the y axis (defaults to 1.2)}

\item{ymin}{y-min of the y axis (defaults to 0)}

\item{pcadims}{for how many principal components shall the dimensional contributions be calculated (note: only 2 can be plotted, defaults to 2)}

\item{studylabel}{you can assign an individual study name to label some of the tables in the relsa object}

\item{severity}{you can assign a prospective severity for your experiment (has no effect on model calculations, just information)}

\item{colorlabel}{you can assign a color label to some tables of the relsa object}

\item{showScree}{shows the scree plot (gets overwritten by showPlot)}

\item{showPlot}{shows the k-means plot (ordered)}

\item{k}{the number of k-clusters in the k-means analysis}
}
\value{
\code{robj} RELSA object - a list with RELSA calculation outputs.
}
\description{
The \code{relsa_wrapper} analyses the data set and calculates various RELSA objects for later analysis.
Please note that the data must be in the RELSA format.
}
