% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surgery.R
\docType{data}
\name{surgery}
\alias{surgery}
\title{Transmitter operated mice}
\format{
A 840 x 10 data frame with surgery data (transmitter implantation).
}
\usage{
data(surgery)
}
\description{
Data from a transmitter implantation experiment with various behavioural data. Also includes sham-operated animals.
Observation time was over 30 days, starting with day=-1 as baseline (physiological state). The treatment columns shows
what operational state the animal is in and the condition field shows two possible types of analgesia. The data comprise
the outcome measures: body weight change (bwc), burrowing overnight (burON), heart rate (hr), heart rate variability (hrv),
temperature (temp) and activity (act).
}
\examples{
data(surgery)
head(surgery)
}
\keyword{datasets}
