% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postop.R
\docType{data}
\name{postop}
\alias{postop}
\title{Post-op data of transmitter-implanted mice}
\format{
A 840 x 16 data frame with post-op data.
}
\usage{
data(postop)
}
\description{
Data from a transmitter implantation experiment with various behavioral
outcome measures, especially short-term indicators such as the Mouse Grimace
Scale and the burrowing parameter. The data also include Sham-operated
animals. The observation time was 30 days, starting with day=-1 as baseline
(physiological state) and day 0 as the intervention day (surgery).
The'treatment' column indicates the surgery subgroup and the condition column
reports the type of analgesia.
}
\examples{
data(postop)
head(postop)
}
\keyword{datasets}
