% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relsa_reference.R
\name{relsa_reference}
\alias{relsa_reference}
\title{RELSA Default Reference Model}
\usage{
relsa_reference()
}
\value{
list with baselines (bsl) and cluster levels (levels)
}
\description{
The \code{relsa_reference} functions calculates the default RELSA reference model - to keeps things simple.
}
