% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surgery.R
\docType{data}
\name{surgery}
\alias{surgery}
\title{Surgery data (transmitter-implanted mice)}
\format{
A 840 x 10 data frame with surgery data (transmitter implantation).
}
\usage{
data(surgery)
}
\description{
Data from a transmitter-implantation experiment with various behavioral
outcome measures. The data also include Sham-operated animals.
The observation time was 30 days, starting with day=-1 as baseline
(physiological state) and day 0 as the intervention day (surgery).
The 'treatment' column indicates the surgery subgroup and the condition
column reports the type of analgesia.
The data comprise the outcome measures: body weight change (bwc), burrowing
overnight (burON), heart rate (hr), heart rate variability (hrv),
temperature (temp) and activity (act).#'
}
\examples{
data(surgery)
head(surgery)
}
\keyword{datasets}
